Implemented yolov4
I've got darknet running at close to the same speed as the C compiled executable. It only loads the neural network once, then processes all the .lowfps video files in order.

I'm getting an "CUDA Error: the launch timed out and was terminated" error, but it's only showing up on my Windows machine. I think it's because I only have 1 GPU to run both the display output and the CUDA libraries. The linux machine that this will be deployed on is headless. So hopefully that error won't pop up there.